feat: Add confirmation dialog for deleting user in UserController@index
This commit adds a confirmation dialog for deleting a user in the `UserController@index` method. When the user clicks on the delete button, a confirmation dialog will be displayed asking if they are sure they want to delete the user. This enhancement improves the user experience by preventing accidental deletions and providing a clear prompt for user confirmation.