```text feat: Fix condition for creating KelompokMetlit in MetlitController@index
The code change fixes the condition for creating a new KelompokMetlit in the MetlitController@index method. The previous condition was checking for the absence of the mahasiswa_id property, but it should actually compare it with the id of the detailMahasiswa object. This fix ensures that the KelompokMetlit is created correctly based on the correct condition.